requirements.txt
langchain==0.1.0
langchain-core==0.1.0
langchain-community==0.0.1
langchain-chroma==0.1.0
langchain-groq==0.1.0
langchain-huggingface==0.0.1
langgraph==0.0.1
python-dotenv==1.0.0
pypdf==3.17.1
chromadb==0.4.0
sentence-transformers==2.2.0
matplotlib==3.8.0
